refactor: 패키지 정리 및 메서드 분리
