+++ src/main/java/container/HandlerAdapter.java
@@ -1,4 +1,6 @@
-package webserver;
+package container;
+
+import container.domain.MappingInfo;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -15,15 +17,20 @@ public class HandlerAdapter {
             throws InvocationTargetException, IllegalAccessException {
         Method method = mappingInfo.getMethod();
         Object object = mappingInfo.getObject();
+
         if (mappingInfo.isStatic()) {
             return mappingInfo.getUrl();
         }
 
         Map<String, String> params = mappingInfo.getParams();
         if (params.isEmpty()) {
-            return TEMPLATES + method.invoke(object);
+            return getResult((String) method.invoke(object));
         } else {
-            return TEMPLATES + method.invoke(object, params.values().toArray(new Object[0]));
+            return getResult((String) method.invoke(object, params.values().toArray(new Object[0])));
         }
+    }
+
+    private static String getResult(String result) {
+        return result.startsWith("/") ? TEMPLATES + result : result;
     }
 }
