+++ src/main/java/container/domain/HttpRequest.java
@@ -1,4 +1,4 @@
-package webserver;
+package container.domain;
 
 import java.util.HashMap;
 import java.util.Map;
