+++ src/main/java/container/domain/MappingInfo.java
@@ -1,4 +1,4 @@
-package webserver;
+package container.domain;
 
 import java.lang.reflect.Method;
 import java.util.Map;
