+++ src/main/java/webserver/WebServer.java
@@ -3,6 +3,7 @@ package webserver;
 import java.net.ServerSocket;
 import java.net.Socket;
 
+import container.RequestHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -11,7 +12,7 @@ public class WebServer {
     private static final int DEFAULT_PORT = 8080;
 
     public static void main(String args[]) throws Exception {
-        int port = 0;
+        int port;
         if (args == null || args.length == 0) {
             port = DEFAULT_PORT;
         } else {
